WIP, pulled out generator code to common
